changed data type for xy_width and xy_height

+++ trsfile/common.py
@@ -106,8 +106,8 @@ class Header(Enum):
 	KEY_LENGTH                       = (0x70, 'KL', False,    int,          4,      0,                 'Key data length in trace data')
 	NUMBER_CHANNELS                  = (0x71, 'CH', False,    int,          4,      0,                 'Number of enabled channels used for acquisition')
 	NUMBER_SCOPES                    = (0x72, 'NO', False,    int,          4,      0,                 'Number of scopes used for acquisition')
-	XY_WIDTH                         = (0x73, 'WI', False,    float,        4,      0,                 'Width in case of XY acquisition')
-	XY_HEIGHT                        = (0x74, 'HE', False,    float,        4,      0,                 'Height in case of XY acquisition')
+	XY_WIDTH                         = (0x73, 'WI', False,    int,          4,      0,                 'Width in case of XY acquisition')
+	XY_HEIGHT                        = (0x74, 'HE', False,    int,          4,      0,                 'Height in case of XY acquisition')
 	XY_NUMBER_MEASUREMENTS_PER_SPOT  = (0x75, 'ME', False,    int,          4,      0,                 'Measurements per spot in case of XY acquisition')
 
 	def __new__(cls, tag, tag_name, is_mandatory, type, length, default, description):
